requirements.txt
ib_insync>=0.9.86
pandas>=2.0
numpy>=1.24
PyYAML>=6.0
aiohttp>=3.9
pytz>=2023.3
pytest>=7.4
